pyproject.toml
[project]
name = "imapbackup"
version = "0.3.0"
description = "Export and backup tool for IMAP mailboxes"
authors = [{ name = "Stefan Schönberger", email = "stefan@sniner.dev" }]
requires-python = ">=3.10"
dependencies = [
    "imapclient",
    "pyyaml",
    "zstandard",
]

[project.optional-dependencies]
graph = ["msal", "httpx"]

[dependency-groups]
dev = [
    "pylint",
    "pytest",
    "imapbackup[graph]",
    "basedpyright>=1.38.1",
]

[tool.uv]
package = true

[tool.pyright]
typeCheckingMode = "standard"
useLibraryCodeForTypes = true
venvPath = "."
venv = ".venv"

[tool.ruff]
line-length = 96
ignore = ["E402"]

[tool.ruff.lint]
per-file-ignores = { "__init__.py" = ["F401"] }

[project.scripts]
ib-mailbox = "imapbackup.cli.mailbox:main"
ib-archive = "imapbackup.cli.archive:main"
ib-copy    = "imapbackup.cli.copy:main"

#[[tool.uv.index]]
#name = "pypi"
#url = "https://pypi.org/simple"

#[tool.uv.sources]
#imapbackup = { path = ".", editable = true }

[tool.pytest.ini_options]
testpaths = ["tests"]
